Allow to provide refresh_token on initialize
In my application I need to support multiple accounts. Not sure if I
missed something, but there is currently no way
to do it. The SDK saves the refresh_token in the SharedPreferences,
assuming that there will always be 1 user account.
If we want to change accounts, we have to logout and then login again.
I would rather be able to switch accounts without
having to logout/login.
In my case I would like to pass the refresh_token to the initialize
method, and it would solve my problems.